requirements.txt
python-roborock>=0.18.0
aiohttp>=3.9.0
pyyaml>=6.0
